add firefox add-on id to manifest.json
so. i tried installing to firefox, but each time i did, it came up with a "this file seems to be corrupt" message. 
but when i try running it as a temporary add-on in firefox (about:debugging#/runtime/this-firefox), it runs just fine... i tried uploading it to mozilla's add-on dev hub privately, (https://addons.mozilla.org/en-US/developers/addons) and it said "Your add-on failed validation with 1 error. The extension ID is required in Manifest Version 3 and above." so uh. yeah

some (very big some) mozilla documentation of id stuff: https://extensionworkshop.com/documentation/develop/extensions-and-the-add-on-id/

sorry if this seems redundant